'ADDedUser'

+++ app/Http/Controllers/Admin/UsersController.php
@@ -8,7 +8,8 @@ use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Role;
 use App\User;
-
+use App\Player;
+use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -16,15 +17,14 @@ class UsersController extends Controller
 {
     public function index()
     {
-
-        $users = User::all();
+        $users = Player::all();
 
         return view('admin.users.index', compact('users'));
     }
 
     public function create()
     {
-
+        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
 
@@ -41,7 +41,7 @@ class UsersController extends Controller
 
     public function edit(User $user)
     {
-
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
 
@@ -60,13 +60,16 @@ class UsersController extends Controller
 
     public function show(User $user)
     {
+        abort_if(Gate::denies('user_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user->load('roles');
+
         return view('admin.users.show', compact('user'));
     }
 
     public function destroy(User $user)
     {
-
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $user->delete();
